refactor: update terminology from "assignments" to "activities" across the application
- Changed text and labels in UI components to reflect the new terminology.
- Updated notification types and messages to use "activity" instead of "assignment".
- Modified database functions and constraints to accommodate the new grading criteria for activities.
- Adjusted tests to ensure they align with the updated terminology and functionality.

+++ test/features/class_records/class_record_calculator_test.dart
@@ -156,7 +156,7 @@ SubmissionModel _submission(DateTime now, {required int score}) {
     userDisplayName: 'Ada Lovelace',
     code: 'answer',
     status: SubmissionStatus.graded,
-    score: score,
+    score: score.toDouble(),
     createdAt: now,
   );
 }
